Confirmation modal

+++ packages/dash-core-components/test/IntegrationTests.py
@@ -4,8 +4,14 @@ import multiprocessing
 import time
 import unittest
 import percy
+import threading
+import platform
+import flask
+import requests
+
 from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
+
 
 class IntegrationTests(unittest.TestCase):
 
@@ -34,10 +40,19 @@ class IntegrationTests(unittest.TestCase):
 
     def tearDown(self):
         time.sleep(3)
-        self.server_process.terminate()
+        if platform.system() == 'Windows':
+            requests.get('http://localhost:8050/stop')
+        else:
+            self.server_process.terminate()
         time.sleep(3)
 
     def startServer(self, app):
+        """
+
+        :param app:
+        :type app: dash.Dash
+        :return:
+        """
         if 'DASH_TEST_PROCESSES' in os.environ:
             processes = int(os.environ['DASH_TEST_PROCESSES'])
         else:
@@ -52,9 +67,32 @@ class IntegrationTests(unittest.TestCase):
                 processes=processes
             )
 
+        def run_windows():
+            app.scripts.config.serve_locally = True
+            app.css.config.serve_locally = True
+
+            @app.server.route('/stop')
+            def _stop_server_windows():
+                stopper = flask.request.environ['werkzeug.server.shutdown']
+                stopper()
+                return 'stop'
+
+            app.run_server(
+                port=8050,
+                debug=False,
+                threaded=True
+            )
+
         # Run on a separate process so that it doesn't block
-        self.server_process = multiprocessing.Process(target=run)
-        self.server_process.start()
+
+        system = platform.system()
+        if system == 'Windows':
+            # multiprocess can't pickle an inner func on windows (closure are not serializable by default on windows)
+            self.server_thread = threading.Thread(target=run_windows)
+            self.server_thread.start()
+        else:
+            self.server_process = multiprocessing.Process(target=run)
+            self.server_process.start()
         time.sleep(5)
 
         # Visit the dash page
